_LEGACY_V1_ARCHIVE/requirements.txt
cryptography
requests
python-dotenv
pandas
numpy
ta
rich
streamlit
watchdog
